tensorboard visualizer algo change of outputs and test fix

+++ tests/enmap-box/enmapbox/apps/SpecDeepMap/test_tensorboard_visualizer.py
@@ -44,27 +44,28 @@ class Test_Tensorboard(TestCase):
 
         result = Processing.runAlgorithm(alg, parameters=io)
 
-        print(result)
-        time.sleep(20)
+        print('results',result)
         # Get the process with the given PID
-        process = psutil.Process(result['PID'])
+        pid = result['PID']
+        print('pid',pid)
+        process_exist = result['Process_exist']
+        process_runs = result['process_runs']
+        print('exist', process_exist)
+        print('runs', process_runs)
 
-        # Assert if the process is running
-        assert process.is_running(), f"Process with PID {result['PID']} is not running."
+        # Assert if the process is not existing or running
+        assert process_exist == True or process_runs == True
+        time.sleep(15)
+        # if process still exist terminate
+        if psutil.pid_exists(result['PID']) == True:
+            process = psutil.Process(result['PID'])
+            # termninate possible childe process and main process
+            for child in process.children(recursive=True):
+              child.kill()
+            process.kill()
 
         # Clean up
 
-        # Kill the process after the check
-
-        # time.sleep(15)   process kills also tensorboard directly after creating. if check if tensorbard gui is open unhashtag this line
-
-        process = psutil.Process(result['PID'])
-        # Recursively kill all child processes
-        for child in process.children(recursive=True):
-            child.kill()  # Force kill child processes
-
-        # Kill the main process
-        process.kill()
         # Remove logg folder
         folder_path_logs_out = BASE_DIR / "test_run" / "lightning_logs"
